# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Нууу очень крутая апишка",
        "description": "Документация для REST API Laravel",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
        "/api/v1/auth/registration": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "badd794a1cc937fb5b851f279bba37e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь успешно зарегестрирован"
                    },
                    "422": {
                        "description": "Неверные данные"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "a7997e7c1e3bfed64d4e56de82415931",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный вход"
                    },
                    "401": {
                        "description": "Неверные учетные данные"
                    },
                    "422": {
                        "description": "Неверные данные"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выход пользователя",
                "description": "Отзывает токен текущего пользователя и завершает сессию авторизации.",
                "operationId": "6def2e57e0a6c223c8d6220642045dfc",
                "responses": {
                    "200": {
                        "description": "Выход выполнен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Выход выполнен успешно"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Получить информацию о текущем пользователе",
                "description": "Возвращает данные авторизованного пользователя по токену.",
                "operationId": "37d33f0281f67d31d05b4c45eea0143b",
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Иван Иванов"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "user@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-10-21T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/email-verify/{user}": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Отправить ссылку для подтверждения email",
                "description": "Создает токен и возвращает ссылку для подтверждения email указанного пользователя.",
                "operationId": "5434379bca9a19468e5a91734063ff0d",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ссылка создана успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ссылка для подтверждения email создана."
                                        },
                                        "verify_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/auth/email-verify?user=1&token=abc123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email уже подтверждён"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        },
        "/api/v1/auth/email-verify": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Подтвердить email по токену",
                "description": "Проверяет токен подтверждения email и устанавливает дату верификации.",
                "operationId": "60e9daea6ed6422a777f3e5f03dcaa9c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user",
                                    "token"
                                ],
                                "properties": {
                                    "user": {
                                        "description": "ID пользователя",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "token": {
                                        "description": "Токен подтверждения email",
                                        "type": "string",
                                        "example": "abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email успешно подтверждён"
                    },
                    "400": {
                        "description": "Неверный или истёкший токен"
                    },
                    "404": {
                        "description": "Запрос подтверждения не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/auth/password/forgot": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Создать токен для сброса пароля",
                "description": "Генерирует токен для сброса пароля и возвращает его (в тестовых целях).",
                "operationId": "8f93622532ddd594ed5e2ac7f626d732",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ссылка для сброса пароля создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ссылка для сброса пароля успешно создана."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "abc123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Неверные данные"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        },
        "/api/v1/auth/password/token-check": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Проверить токен сброса пароля",
                "description": "Проверяет, действителен ли токен, созданный для сброса пароля.",
                "operationId": "d82c25a24c7b3f45239c12e5b56df4e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Токен действителен"
                    },
                    "400": {
                        "description": "Неверный или истёкший токен"
                    },
                    "404": {
                        "description": "Запрос не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/auth/password/reset": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Сбросить пароль пользователя",
                "description": "Позволяет задать новый пароль при наличии действительного токена.",
                "operationId": "8d87fd78d1bf7ed14ae8a41a0d3f75ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "abc123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newPassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пароль успешно обновлён"
                    },
                    "400": {
                        "description": "Неверный токен или email"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}